CT-887 move component/configurations tests out of common
