+++ tests/Components/ConfigurationRowTest.php
@@ -1,5 +1,5 @@
 <?php
-namespace Keboola\Test\Common;
+namespace Keboola\Test\Components;
 
 use Keboola\StorageApi\Client;
 use Keboola\StorageApi\ClientException;
